requirements.txt
google-generativeai
sentence-transformers
faiss-cpu
numpy
pandas
python-dotenv
fastapi
uvicorn
